+++ wandb/sdk_py27/__init__.py
@@ -1,4 +1,4 @@
-#
+# File is generated by: tox -e codemod
 # -*- coding: utf-8 -*-
 """
 module sdk
