+++ wandb/sdk_py27/wandb_alerts.py
@@ -1,4 +1,4 @@
-#
+# File is generated by: tox -e codemod
 from enum import Enum
 
 """
