+++ wandb/sdk_py27/wandb_artifacts.py
@@ -1,4 +1,4 @@
-#
+# File is generated by: tox -e codemod
 import contextlib
 import re
 import os
@@ -37,9 +37,9 @@ _REQUEST_POOL_MAXSIZE = 64
 class Artifact(object):
     """An artifact object you can write files into, and pass to log_artifact."""
 
-    name: str
-    description: Optional[str]
-    distributed_id: Optional[str]
+    # name: str
+    # description: Optional[str]
+    # distributed_id: Optional[str]
 
     def __init__(self, name, type, description=None, metadata=None):
         if not re.match(r"^[a-zA-Z0-9_\-.]+$", name):
