+++ wandb/sdk_py27/wandb_config.py
@@ -1,4 +1,4 @@
-#
+# File is generated by: tox -e codemod
 # -*- coding: utf-8 -*-
 """
 config.
@@ -122,7 +122,7 @@ class Config(object):
     def __getitem__(self, key):
         return self._items[key]
 
-    def _check_locked(self, key, ignore_locked=False) -> bool:
+    def _check_locked(self, key, ignore_locked=False):
         locked = self._locked.get(key)
         if locked is not None:
             locked_user = self._users_inv[locked]
@@ -210,7 +210,7 @@ class Config(object):
             self.update(conf_dict)
 
     def _sanitize_dict(
-        self, config_dict, allow_val_change=None, ignore_keys: set = None
+        self, config_dict, allow_val_change=None, ignore_keys = None
     ):
         sanitized = {}
         for k, v in six.iteritems(config_dict):
