+++ wandb/sdk_py27/wandb_helper.py
@@ -1,4 +1,4 @@
-#
+# File is generated by: tox -e codemod
 import inspect
 import types
