+++ wandb/sdk_py27/wandb_init.py
@@ -1,4 +1,4 @@
-#
+# File is generated by: tox -e codemod
 # -*- coding: utf-8 -*-
 """
 wandb.init() indicates the beginning of a new run. In an ML training pipeline,
@@ -70,7 +70,7 @@ class _WandbInit(object):
         self._wl = None
         self._reporter = None
 
-    def setup(self, kwargs) -> None:
+    def setup(self, kwargs):
         """
         Complete setup for wandb.init(). This includes parsing all arguments,
         applying them with settings and enabling logging.
@@ -82,7 +82,7 @@ class _WandbInit(object):
         _set_logger(self._wl._get_logger())
 
         # Start with settings from wandb library singleton
-        settings: Settings = self._wl._clone_settings()
+        settings = self._wl._clone_settings()
         settings_param = kwargs.pop("settings", None)
         if settings_param:
             settings._apply_settings(settings_param)
@@ -91,7 +91,7 @@ class _WandbInit(object):
             settings=settings.duplicate().freeze()
         )
 
-        sm_config: Dict = {} if settings.sagemaker_disable else sagemaker.parse_sm_config()
+        sm_config = {} if settings.sagemaker_disable else sagemaker.parse_sm_config()
         if sm_config:
             sm_api_key = sm_config.get("wandb_api_key", None)
             sm_run, sm_env = sagemaker.parse_sm_resources()
@@ -319,7 +319,7 @@ class _WandbInit(object):
         logger.info("Logging user logs to {}".format(settings.log_user))
         logger.info("Logging internal logs to {}".format(settings.log_internal))
 
-    def init(self) -> Union[Run, Dummy, None]:  # noqa: C901
+    def init(self):  # noqa: C901
         trigger.call("on_init", **self.kwargs)
         s = self.settings
         sweep_config = self.sweep_config
@@ -498,31 +498,31 @@ def getcaller():
 
 
 def init(
-    job_type: Optional[str] = None,
+    job_type = None,
     dir=None,
-    config: Union[Dict, str, None] = None,
-    project: Optional[str] = None,
-    entity: Optional[str] = None,
-    reinit: bool = None,
-    tags: Optional[Sequence] = None,
-    group: Optional[str] = None,
-    name: Optional[str] = None,
-    notes: Optional[str] = None,
-    magic: Union[dict, str, bool] = None,
+    config = None,
+    project = None,
+    entity = None,
+    reinit = None,
+    tags = None,
+    group = None,
+    name = None,
+    notes = None,
+    magic = None,
     config_exclude_keys=None,
     config_include_keys=None,
-    anonymous: Optional[str] = None,
-    mode: Optional[str] = None,
-    allow_val_change: Optional[bool] = None,
-    resume: Optional[Union[bool, str]] = None,
-    force: Optional[bool] = None,
+    anonymous = None,
+    mode = None,
+    allow_val_change = None,
+    resume = None,
+    force = None,
     tensorboard=None,  # alias for sync_tensorboard
     sync_tensorboard=None,
     monitor_gym=None,
     save_code=None,
     id=None,
-    settings: Union[Settings, Dict[str, Any], None] = None,
-) -> Union[Run, Dummy, None]:
+    settings = None,
+):
     """
     Start a new tracked run with `wandb.init()`.
