+++ wandb/sdk_py27/wandb_login.py
@@ -1,4 +1,4 @@
-#
+# File is generated by: tox -e codemod
 # -*- coding: utf-8 -*-
 """
 Log in to Weights & Biases, authenticating your machine to log data to your
@@ -45,8 +45,8 @@ def login(anonymous=None, key=None, relogin=None, host=None, force=None):
 
 class _WandbLogin(object):
     def __init__(self):
-        self.kwargs: Optional[Dict] = None
-        self._settings: Optional[Settings] = None
+        self.kwargs = None
+        self._settings = None
         self._backend = None
         self._silent = None
         self._wl = None
@@ -56,7 +56,7 @@ class _WandbLogin(object):
         self.kwargs = kwargs
 
         # built up login settings
-        login_settings: Settings = wandb.Settings()
+        login_settings = wandb.Settings()
         settings_param = kwargs.pop("_settings", None)
         if settings_param:
             login_settings._apply_settings(settings_param)
@@ -124,7 +124,7 @@ class _WandbLogin(object):
 
     def update_session(self, key):
         _logger = wandb.setup()._get_logger()
-        settings: Settings = wandb.Settings()
+        settings = wandb.Settings()
         login_settings = dict(api_key=key) if key else dict(mode="offline")
         settings._apply_source_login(login_settings, _logger=_logger)
         self._wl._update(settings=settings)
