+++ wandb/sdk_py27/wandb_run.py
@@ -1,4 +1,4 @@
-#
+# File is generated by: tox -e codemod
 # -*- coding: utf-8 -*-
 
 from __future__ import print_function
@@ -86,18 +86,18 @@ RUN_NAME_COLOR = "#cdcd00"
 
 class ExitHooks(object):
 
-    exception: Optional[BaseException] = None
+    exception = None
 
-    def __init__(self) -> None:
+    def __init__(self):
         self.exit_code = 0
         self.exception = None
 
-    def hook(self) -> None:
+    def hook(self):
         self._orig_exit = sys.exit
         sys.exit = self.exit
         sys.excepthook = self.exc_handler
 
-    def exit(self, code: object = 0) -> NoReturn:
+    def exit(self, code = 0):
         orig_code = code
         if code is None:
             code = 0
@@ -106,12 +106,12 @@ class ExitHooks(object):
         self.exit_code = code
         self._orig_exit(orig_code)
 
-    def was_ctrl_c(self) -> bool:
+    def was_ctrl_c(self):
         return isinstance(self.exception, KeyboardInterrupt)
 
     def exc_handler(
-        self, exc_type: Type[BaseException], exc: BaseException, tb: TracebackType
-    ) -> None:
+        self, exc_type, exc, tb
+    ):
         self.exit_code = 1
         self.exception = exc
         if issubclass(exc_type, Error):
@@ -129,7 +129,7 @@ class RunStatusChecker(object):
     For now, we just use this to figure out if the user has requested a stop.
     """
 
-    def __init__(self, interface: BackendSender, polling_interval: int = 15) -> None:
+    def __init__(self, interface, polling_interval = 15):
         self._interface = interface
         self._polling_interval = polling_interval
 
@@ -138,7 +138,7 @@ class RunStatusChecker(object):
         self._thread.daemon = True
         self._thread.start()
 
-    def check_status(self) -> None:
+    def check_status(self):
         join_requested = False
         while not join_requested:
             status_response = self._interface.communicate_status(check_stop_req=True)
@@ -150,10 +150,10 @@ class RunStatusChecker(object):
                     return
             join_requested = self._join_event.wait(self._polling_interval)
 
-    def stop(self) -> None:
+    def stop(self):
         self._join_event.set()
 
-    def join(self) -> None:
+    def join(self):
         self.stop()
         self._thread.join()
 
@@ -179,51 +179,51 @@ class Run(object):
             final value.
     """
 
-    _telemetry_obj: telemetry.TelemetryRecord
-    _teardown_hooks: List[Callable[[], None]]
-    _tags: Optional[Tuple[Any, ...]]
+    # _telemetry_obj: telemetry.TelemetryRecord
+    # _teardown_hooks: List[Callable[[], None]]
+    # _tags: Optional[Tuple[Any, ...]]
 
-    _entity: Optional[str]
-    _project: Optional[str]
-    _group: Optional[str]
-    _job_type: Optional[str]
-    _name: Optional[str]
-    _notes: Optional[str]
+    # _entity: Optional[str]
+    # _project: Optional[str]
+    # _group: Optional[str]
+    # _job_type: Optional[str]
+    # _name: Optional[str]
+    # _notes: Optional[str]
 
-    _run_obj: Optional[RunRecord]
-    _run_obj_offline: Optional[RunRecord]
+    # _run_obj: Optional[RunRecord]
+    # _run_obj_offline: Optional[RunRecord]
     # Use string literal anotation because of type reference loop
-    _backend: Optional["wandb.sdk.backend.backend.Backend"]
-    _wl: Optional[_WandbSetup]
+    # _backend: Optional["wandb.sdk.backend.backend.Backend"]
+    # _wl: Optional[_WandbSetup]
 
-    _upgraded_version_message: Optional[str]
-    _deleted_version_message: Optional[str]
-    _yanked_version_message: Optional[str]
+    # _upgraded_version_message: Optional[str]
+    # _deleted_version_message: Optional[str]
+    # _yanked_version_message: Optional[str]
 
-    _out_redir: Optional[redirect.RedirectBase]
-    _err_redir: Optional[redirect.RedirectBase]
-    _redirect_cb: Optional[Callable[[str, str], None]]
-    _output_writer: Optional["WriteSerializingFile"]
+    # _out_redir: Optional[redirect.RedirectBase]
+    # _err_redir: Optional[redirect.RedirectBase]
+    # _redirect_cb: Optional[Callable[[str, str], None]]
+    # _output_writer: Optional["WriteSerializingFile"]
 
-    _atexit_cleanup_called: bool
-    _hooks: Optional[ExitHooks]
-    _exit_code: Optional[int]
+    # _atexit_cleanup_called: bool
+    # _hooks: Optional[ExitHooks]
+    # _exit_code: Optional[int]
 
-    _run_status_checker: Optional[RunStatusChecker]
-    _poll_exit_response: Optional[PollExitResponse]
+    # _run_status_checker: Optional[RunStatusChecker]
+    # _poll_exit_response: Optional[PollExitResponse]
 
-    _sampled_history: Optional[Dict[str, Union[List[int], List[float]]]]
+    # _sampled_history: Optional[Dict[str, Union[List[int], List[float]]]]
 
-    _use_redirect: bool
-    _stdout_slave_fd: Optional[int]
-    _stderr_slave_fd: Optional[int]
+    # _use_redirect: bool
+    # _stdout_slave_fd: Optional[int]
+    # _stderr_slave_fd: Optional[int]
 
     def __init__(
         self,
-        settings: Settings,
-        config: Optional[Dict[str, Any]] = None,
-        sweep_config: Optional[Dict[str, Any]] = None,
-    ) -> None:
+        settings,
+        config = None,
+        sweep_config = None,
+    ):
         self._config = wandb_config.Config()
         self._config._set_callback(self._config_callback)
         self._config._set_settings(settings)
@@ -239,7 +239,7 @@ class Run(object):
 
         self._settings = settings
         self._wl = None
-        self._reporter: Optional[Reporter] = None
+        self._reporter = None
 
         self._entity = None
         self._project = None
@@ -318,18 +318,18 @@ class Run(object):
         self._use_redirect = True
         self._progress_step = 0
 
-    def _telemetry_callback(self, telem_obj: telemetry.TelemetryRecord) -> None:
+    def _telemetry_callback(self, telem_obj):
         self._telemetry_obj.MergeFrom(telem_obj)
 
-    def _freeze(self) -> None:
+    def _freeze(self):
         self._frozen = True
 
-    def __setattr__(self, attr: str, value: object) -> None:
+    def __setattr__(self, attr, value):
         if getattr(self, "_frozen", None) and not hasattr(self, attr):
             raise Exception("Attribute {} is not supported on Run object.".format(attr))
         super(Run, self).__setattr__(attr, value)
 
-    def _telemetry_imports(self, imp: telemetry.TelemetryImports) -> None:
+    def _telemetry_imports(self, imp):
         mods = sys.modules
         if mods.get("torch"):
             imp.torch = True
@@ -354,7 +354,7 @@ class Run(object):
         if mods.get("transformers"):
             imp.transformers = True
 
-    def _init_from_settings(self, settings: Settings) -> None:
+    def _init_from_settings(self, settings):
         if settings.entity is not None:
             self._entity = settings.entity
         if settings.project is not None:
@@ -370,7 +370,7 @@ class Run(object):
         if settings.run_tags is not None:
             self._tags = settings.run_tags
 
-    def _make_proto_run(self, run: RunRecord) -> None:
+    def _make_proto_run(self, run):
         """Populate protocol buffer RunData for interface/interface."""
         if self._entity is not None:
             run.entity = self._entity
@@ -393,14 +393,14 @@ class Run(object):
             run.start_time.FromSeconds(int(self._start_time))
         # Note: run.config is set in interface/interface:_make_run()
 
-    def __getstate__(self) -> None:
+    def __getstate__(self):
         pass
 
-    def __setstate__(self, state: Any) -> None:
+    def __setstate__(self, state):
         pass
 
     @property
-    def dir(self) -> str:
+    def dir(self):
         """
         Returns:
             (str): The directory where all of the files associated with the run are
@@ -409,7 +409,7 @@ class Run(object):
         return self._settings.files_dir
 
     @property
-    def config(self) -> wandb_config.Config:
+    def config(self):
         """
         Returns:
             (Config): A config object (similar to a nested dict) of key
@@ -418,11 +418,11 @@ class Run(object):
         return self._config
 
     @property
-    def config_static(self) -> wandb_config.ConfigStatic:
+    def config_static(self):
         return wandb_config.ConfigStatic(self._config)
 
     @property
-    def name(self) -> Optional[str]:
+    def name(self):
         """
         Returns:
             (str): the display name of the run. It does not need to be unique
@@ -435,13 +435,13 @@ class Run(object):
         return self._run_obj.display_name
 
     @name.setter
-    def name(self, name: str) -> None:
+    def name(self, name):
         self._name = name
         if self._backend:
             self._backend.interface.publish_run(self)
 
     @property
-    def notes(self) -> Optional[str]:
+    def notes(self):
         r"""
         Returns:
             (str): notes associated with the run. Notes can be a multiline string
@@ -453,13 +453,13 @@ class Run(object):
         return self._run_obj.notes
 
     @notes.setter
-    def notes(self, notes: str) -> None:
+    def notes(self, notes):
         self._notes = notes
         if self._backend:
             self._backend.interface.publish_run(self)
 
     @property
-    def tags(self) -> Optional[Tuple]:
+    def tags(self):
         """
         Returns:
             (Tuple[str]): tags associated with the run
@@ -472,13 +472,13 @@ class Run(object):
         return None
 
     @tags.setter
-    def tags(self, tags: Sequence) -> None:
+    def tags(self, tags):
         self._tags = tuple(tags)
         if self._backend:
             self._backend.interface.publish_run(self)
 
     @property
-    def id(self) -> str:
+    def id(self):
         """id property.
 
         Returns:
@@ -487,7 +487,7 @@ class Run(object):
         return self._run_id
 
     @property
-    def sweep_id(self) -> Optional[str]:
+    def sweep_id(self):
         """
         Returns:
             (str, optional): the sweep id associated with the run or None
@@ -497,7 +497,7 @@ class Run(object):
         return self._run_obj.sweep_id or None
 
     @property
-    def path(self) -> str:
+    def path(self):
         """
         Returns:
             (str): the path to the run `[entity]/[project]/[run_id]`
@@ -509,7 +509,7 @@ class Run(object):
         return "/".join(parts)
 
     @property
-    def start_time(self) -> float:
+    def start_time(self):
         """
         Returns:
             (int): the unix time stamp in seconds when the run started
@@ -520,7 +520,7 @@ class Run(object):
             return self._run_obj.start_time.ToSeconds()
 
     @property
-    def starting_step(self) -> int:
+    def starting_step(self):
         """
         Returns:
             (int): the first step of the run
@@ -531,7 +531,7 @@ class Run(object):
             return self._run_obj.starting_step
 
     @property
-    def resumed(self) -> bool:
+    def resumed(self):
         """
         Returns:
             (bool): whether or not the run was resumed
@@ -541,7 +541,7 @@ class Run(object):
         return False
 
     @property
-    def step(self) -> int:
+    def step(self):
         """
         Every time you call wandb.log() it will by default increment the step
         counter.
@@ -551,25 +551,25 @@ class Run(object):
         """
         return self.history._step
 
-    def project_name(self) -> str:
+    def project_name(self):
         run_obj = self._run_obj or self._run_obj_offline
         return run_obj.project if run_obj else ""
 
     @property
-    def mode(self) -> str:
+    def mode(self):
         """For compatibility with `0.9.x` and earlier, deprecate eventually."""
         return "dryrun" if self._settings._offline else "run"
 
     @property
-    def offline(self) -> bool:
+    def offline(self):
         return self._settings._offline
 
     @property
-    def disabled(self) -> bool:
+    def disabled(self):
         return self._settings._noop
 
     @property
-    def group(self) -> str:
+    def group(self):
         """
         Setting a group helps the W&B UI organize runs in a sensible way.
 
@@ -585,19 +585,19 @@ class Run(object):
         return run_obj.run_group if run_obj else ""
 
     @property
-    def job_type(self) -> str:
+    def job_type(self):
         run_obj = self._run_obj or self._run_obj_offline
         return run_obj.job_type if run_obj else ""
 
     @property
-    def project(self) -> str:
+    def project(self):
         """
         Returns:
             (str): name of W&B project associated with run.
         """
         return self.project_name()
 
-    def get_url(self) -> Optional[str]:
+    def get_url(self):
         """
         Returns:
             (str, optional): url for the W&B run or None if the run
@@ -608,7 +608,7 @@ class Run(object):
             return None
         return self._get_run_url()
 
-    def get_project_url(self) -> Optional[str]:
+    def get_project_url(self):
         """
         Returns:
             (str, optional): url for the W&B project associated with
@@ -619,7 +619,7 @@ class Run(object):
             return None
         return self._get_project_url()
 
-    def get_sweep_url(self) -> Optional[str]:
+    def get_sweep_url(self):
         """
         Returns:
             (str, optional): url for the sweep associated with the run
@@ -631,7 +631,7 @@ class Run(object):
         return self._get_sweep_url()
 
     @property
-    def url(self) -> Optional[str]:
+    def url(self):
         """
         Returns:
             (str): name of W&B url associated with run.
@@ -639,7 +639,7 @@ class Run(object):
         return self.get_url()
 
     @property
-    def entity(self) -> str:
+    def entity(self):
         """
         Returns:
             (str): name of W&B entity associated with run. Entity is either
@@ -648,8 +648,8 @@ class Run(object):
         return self._entity or ""
 
     def _repr_mimebundle_(
-        self, include: Any = None, exclude: Any = None
-    ) -> Dict[str, str]:
+        self, include = None, exclude = None
+    ):
         url = self._get_run_url()
         style = "border:none;width:100%;height:400px"
         s = '<h1>Run({})</h1><iframe src="{}" style="{}"></iframe>'.format(
@@ -658,31 +658,31 @@ class Run(object):
         return {"text/html": s}
 
     def _config_callback(
-        self, key: str = None, val: object = None, data: Dict[str, object] = None
-    ) -> None:
+        self, key = None, val = None, data = None
+    ):
         logger.info("config_cb %s %s %s", key, val, data)
         if not self._backend or not self._backend.interface:
             return
         self._backend.interface.publish_config(key=key, val=val, data=data)
 
-    def _summary_update_callback(self, summary_record: SummaryRecord) -> None:
+    def _summary_update_callback(self, summary_record):
         if self._backend:
             self._backend.interface.publish_summary(summary_record)
 
-    def _summary_get_current_summary_callback(self) -> Dict[str, Any]:
+    def _summary_get_current_summary_callback(self):
         if not self._backend:
             return {}
         ret = self._backend.interface.communicate_summary()
         return proto_util.dict_from_proto_list(ret.item)
 
-    def _datatypes_callback(self, fname: str) -> None:
+    def _datatypes_callback(self, fname):
         if not self._backend:
             return
         files = dict(files=[(fname, "now")])
         self._backend.interface.publish_files(files)
 
     # TODO(jhr): codemod add: PEP 3102 -- Keyword-Only Arguments
-    def _history_callback(self, row: Dict[str, Any], step: int) -> None:
+    def _history_callback(self, row, step):
 
         # TODO(jhr): move visualize hack somewhere else
         visualize_persist_config = False
@@ -720,32 +720,32 @@ class Run(object):
         if self._backend:
             self._backend.interface.publish_history(row, step)
 
-    def _console_callback(self, name: str, data: str) -> None:
+    def _console_callback(self, name, data):
         # logger.info("console callback: %s, %s", name, data)
         if self._backend:
             self._backend.interface.publish_output(name, data)
 
     def _tensorboard_callback(
-        self, logdir: str, save: bool = None, root_logdir: str = None
-    ) -> None:
+        self, logdir, save = None, root_logdir = None
+    ):
         logger.info("tensorboard callback: %s, %s", logdir, save)
         save = True if save is None else save
         if self._backend:
             self._backend.interface.publish_tbdata(logdir, save, root_logdir)
 
-    def _set_library(self, library: _WandbSetup) -> None:
+    def _set_library(self, library):
         self._wl = library
 
-    def _set_backend(self, backend: "wandb.sdk.backend.backend.Backend") -> None:
+    def _set_backend(self, backend):
         self._backend = backend
 
-    def _set_reporter(self, reporter: Reporter) -> None:
+    def _set_reporter(self, reporter):
         self._reporter = reporter
 
-    def _set_teardown_hooks(self, hooks: List[Callable[[], None]]) -> None:
+    def _set_teardown_hooks(self, hooks):
         self._teardown_hooks = hooks
 
-    def _set_run_obj(self, run_obj: RunRecord) -> None:
+    def _set_run_obj(self, run_obj):
         self._run_obj = run_obj
         self._entity = run_obj.entity
         self._project = run_obj.project
@@ -773,12 +773,12 @@ class Run(object):
             url=self._get_run_url(),
         )
 
-    def _set_run_obj_offline(self, run_obj: RunRecord) -> None:
+    def _set_run_obj_offline(self, run_obj):
         self._run_obj_offline = run_obj
 
     def _add_singleton(
-        self, data_type: str, key: str, value: Dict[Union[int, str], str]
-    ) -> None:
+        self, data_type, key, value
+    ):
         """Stores a singleton item to wandb config.
 
         A singleton in this context is a piece of data that is continually
@@ -806,11 +806,11 @@ class Run(object):
 
     def log(
         self,
-        data: Dict[str, Any],
-        step: int = None,
-        commit: bool = None,
-        sync: bool = None,
-    ) -> None:
+        data,
+        step = None,
+        commit = None,
+        sync = None,
+    ):
         """Log a dict to the global run's history.
 
         `wandb.log` can be used to log everything from scalars to histograms, media
@@ -947,10 +947,10 @@ class Run(object):
 
     def save(
         self,
-        glob_str: Optional[str] = None,
-        base_path: Optional[str] = None,
-        policy: str = "live",
-    ) -> Union[bool, List[str]]:
+        glob_str = None,
+        base_path = None,
+        policy = "live",
+    ):
         """ Ensure all files matching *glob_str* are synced to wandb with the policy specified.
 
         Arguments:
@@ -1038,14 +1038,14 @@ class Run(object):
 
     def restore(
         self,
-        name: str,
-        run_path: Optional[str] = None,
-        replace: bool = False,
-        root: Optional[str] = None,
-    ) -> Union[None, TextIO]:
+        name,
+        run_path = None,
+        replace = False,
+        root = None,
+    ):
         return restore(name, run_path or self.path, replace, root or self.dir)
 
-    def finish(self, exit_code: int = None) -> None:
+    def finish(self, exit_code = None):
         """Marks a run as finished, and finishes uploading all data.  This is
         used when creating multiple runs in the same process.  We automatically
         call this method when your script exits.
@@ -1061,7 +1061,7 @@ class Run(object):
             self._wl._global_run_stack.pop()
         module.unset_globals()
 
-    def join(self, exit_code: int = None) -> None:
+    def join(self, exit_code = None):
         """Deprecated alias for `finish()` - please use finish"""
         self.finish(exit_code=exit_code)
 
@@ -1083,18 +1083,18 @@ class Run(object):
         )
         return visualization
 
-    def _set_upgraded_version_message(self, msg: str) -> None:
+    def _set_upgraded_version_message(self, msg):
         self._upgraded_version_message = msg
 
-    def _set_deleted_version_message(self, msg: str) -> None:
+    def _set_deleted_version_message(self, msg):
         self._deleted_version_message = msg
 
-    def _set_yanked_version_message(self, msg: str) -> None:
+    def _set_yanked_version_message(self, msg):
         self._yanked_version_message = msg
 
     def _add_panel(
-        self, visualize_key: str, panel_type: str, panel_config: dict
-    ) -> None:
+        self, visualize_key, panel_type, panel_config
+    ):
         if "visualize" not in self._config["_wandb"]:
             self._config["_wandb"]["visualize"] = dict()
         self._config["_wandb"]["visualize"][visualize_key] = {
@@ -1104,7 +1104,7 @@ class Run(object):
 
         self._config_callback(data=self._config._as_dict())
 
-    def _get_url_query_string(self) -> str:
+    def _get_url_query_string(self):
         s = self._settings
 
         # TODO(jhr): migrate to new settings, but for now this is safer
@@ -1115,7 +1115,7 @@ class Run(object):
         api_key = apikey.api_key(settings=s)
         return "?" + urlencode({"apiKey": api_key})
 
-    def _get_project_url(self) -> str:
+    def _get_project_url(self):
         s = self._settings
         r = self._run_obj
         if not r:
@@ -1127,7 +1127,7 @@ class Run(object):
         )
         return url
 
-    def _get_run_url(self) -> str:
+    def _get_run_url(self):
         s = self._settings
         r = self._run_obj
         if not r:
@@ -1139,7 +1139,7 @@ class Run(object):
         )
         return url
 
-    def _get_sweep_url(self) -> str:
+    def _get_sweep_url(self):
         """Generate a url for a sweep.
 
         Returns:
@@ -1165,13 +1165,13 @@ class Run(object):
             qs=qs,
         )
 
-    def _get_run_name(self) -> str:
+    def _get_run_name(self):
         r = self._run_obj
         if not r:
             return ""
         return r.display_name
 
-    def _display_run(self) -> None:
+    def _display_run(self):
         project_url = self._get_project_url()
         run_url = self._get_run_url()
         sweep_url = self._get_sweep_url()
@@ -1253,16 +1253,16 @@ class Run(object):
 
     def _redirect(
         self,
-        stdout_slave_fd: Optional[int],
-        stderr_slave_fd: Optional[int],
-        console: SettingsConsole = None,
-    ) -> None:
+        stdout_slave_fd,
+        stderr_slave_fd,
+        console = None,
+    ):
         if console is None:
             console = self._settings._console
         logger.info("redirect: %s", console)
 
-        out_redir: redirect.RedirectBase
-        err_redir: redirect.RedirectBase
+        # out_redir: redirect.RedirectBase
+        # err_redir: redirect.RedirectBase
         if console == self._settings.Console.REDIRECT:
             logger.info("Redirecting console.")
             out_cap = redirect.Capture(
@@ -1279,7 +1279,7 @@ class Run(object):
             )
             if os.name == "nt":
 
-                def wrap_fallback() -> None:
+                def wrap_fallback():
                     if self._out_redir:
                         self._out_redir.uninstall()
                     if self._err_redir:
@@ -1337,7 +1337,7 @@ class Run(object):
         #     self.stderr_redirector = stderr_redirector
         # logger.info("redirect done")
 
-    def _restore(self) -> None:
+    def _restore(self):
         logger.info("restore")
         # TODO(jhr): drain and shutdown all threads
         if self._use_redirect:
@@ -1357,7 +1357,7 @@ class Run(object):
             sys.stderr = self._save_stderr
         logger.info("restore done")
 
-    def _atexit_cleanup(self, exit_code: int = None) -> None:
+    def _atexit_cleanup(self, exit_code = None):
         if self._backend is None:
             logger.warning("process exited without backend configured")
             return
@@ -1395,7 +1395,7 @@ class Run(object):
                 return
             self._on_final()
 
-    def _console_start(self) -> None:
+    def _console_start(self):
         logger.info("atexit reg")
         self._hooks = ExitHooks()
         self._hooks.hook()
@@ -1409,16 +1409,16 @@ class Run(object):
         self._output_writer = WriteSerializingFile(open(output_log_path, "wb"))
         self._redirect(self._stdout_slave_fd, self._stderr_slave_fd)
 
-    def _console_stop(self) -> None:
+    def _console_stop(self):
         self._restore()
         if self._output_writer:
             self._output_writer.close()
             self._output_writer = None
 
-    def _on_init(self) -> None:
+    def _on_init(self):
         self._show_version_info()
 
-    def _on_start(self) -> None:
+    def _on_start(self):
         # TODO: make offline mode in jupyter use HTML
         if self._settings._offline:
             wandb.termlog("Offline run mode, not syncing to the cloud.")
@@ -1438,10 +1438,10 @@ class Run(object):
 
     def _pusher_print_status(
         self,
-        progress: FilePusherStats,
-        prefix: bool = True,
-        done: Optional[bool] = False,
-    ) -> None:
+        progress,
+        prefix = True,
+        done = False,
+    ):
         if self._settings._offline:
             return
 
@@ -1452,7 +1452,7 @@ class Run(object):
         )
 
         if self._jupyter_progress:
-            percent_done: float
+            # percent_done: float
             if progress.total_bytes == 0:
                 percent_done = 1
             else:
@@ -1482,10 +1482,10 @@ class Run(object):
                 # clear progress line.
                 wandb.termlog(" " * 79, prefix=prefix)
 
-    def _on_finish_progress(self, progress: FilePusherStats, done: bool = None) -> None:
+    def _on_finish_progress(self, progress, done = None):
         self._pusher_print_status(progress, done=done)
 
-    def _wait_for_finish(self) -> PollExitResponse:
+    def _wait_for_finish(self):
         while True:
             if self._backend:
                 poll_exit_resp = self._backend.interface.communicate_poll_exit()
@@ -1500,7 +1500,7 @@ class Run(object):
                     return poll_exit_resp
             time.sleep(2)
 
-    def _on_finish(self) -> None:
+    def _on_finish(self):
         trigger.call("on_finished")
 
         # populate final import telemetry
@@ -1557,7 +1557,7 @@ class Run(object):
         if self._run_status_checker:
             self._run_status_checker.join()
 
-    def _on_final(self) -> None:
+    def _on_final(self):
         # check for warnings and errors, show log file locations
         if self._reporter:
             # TODO: handle warnings and errors nicely in jupyter
@@ -1629,7 +1629,7 @@ class Run(object):
 
         self._show_version_info(footer=True)
 
-    def _show_version_info(self, footer: bool = None) -> None:
+    def _show_version_info(self, footer = None):
         package_problem = False
         if self._deleted_version_message:
             wandb.termerror(self._deleted_version_message)
@@ -1642,7 +1642,7 @@ class Run(object):
             if self._upgraded_version_message:
                 wandb.termlog(self._upgraded_version_message)
 
-    def _show_summary(self) -> None:
+    def _show_summary(self):
         if self._final_summary:
             logger.info("rendering summary")
             max_len = max([len(k) for k in self._final_summary.keys()])
@@ -1671,7 +1671,7 @@ class Run(object):
                 wandb.termlog("Run summary:")
                 wandb.termlog(summary_lines)
 
-    def _show_history(self) -> None:
+    def _show_history(self):
         if not self._sampled_history:
             return
 
@@ -1705,7 +1705,7 @@ class Run(object):
                 history_lines += format_str.format(*row)
             wandb.termlog(history_lines)
 
-    def _show_files(self) -> None:
+    def _show_files(self):
         if not self._poll_exit_response or not self._poll_exit_response.file_counts:
             return
         if self._settings._offline:
@@ -1727,7 +1727,7 @@ class Run(object):
         else:
             wandb.termlog(file_str)
 
-    def _save_job_spec(self) -> None:
+    def _save_job_spec(self):
         envdict = dict(python="python3.6", requirements=[],)
         varsdict = {"WANDB_DISABLE_CODE": "True"}
         source = dict(
@@ -1758,7 +1758,7 @@ class Run(object):
         self.save(spec_filename)
 
     # TODO(jhr): annotate this
-    def watch(self, models, criterion=None, log="gradients", log_freq=100, idx=None) -> None:  # type: ignore
+    def watch(self, models, criterion=None, log="gradients", log_freq=100, idx=None):  # type: ignore
         wandb.watch(models, criterion, log, log_freq, idx)
 
     # TODO(jhr): annotate this
@@ -1818,11 +1818,11 @@ class Run(object):
     # TODO(jhr): annotate this
     def log_artifact(
         self,
-        artifact_or_path: Union[wandb_artifacts.Artifact, str],
-        name: Optional[str] = None,
-        type: Optional[str] = None,
-        aliases: Optional[List[str]] = None,
-    ) -> wandb_artifacts.Artifact:
+        artifact_or_path,
+        name = None,
+        type = None,
+        aliases = None,
+    ):
         """ Declare an artifact as output of a run.
 
         Arguments:
@@ -1851,12 +1851,12 @@ class Run(object):
 
     def upsert_artifact(
         self,
-        artifact_or_path: Union[wandb_artifacts.Artifact, str],
-        name: Optional[str] = None,
-        type: Optional[str] = None,
-        aliases: Optional[List[str]] = None,
-        distributed_id: Optional[str] = None,
-    ) -> wandb_artifacts.Artifact:
+        artifact_or_path,
+        name = None,
+        type = None,
+        aliases = None,
+        distributed_id = None,
+    ):
         """ Declare (or append tp) a non-finalized artifact as output of a run. Note that you must call
         run.finish_artifact() to finalize the artifact. This is useful when distributed jobs
         need to all contribute to the same artifact.
@@ -1902,12 +1902,12 @@ class Run(object):
 
     def finish_artifact(
         self,
-        artifact_or_path: Union[wandb_artifacts.Artifact, str],
-        name: Optional[str] = None,
-        type: Optional[str] = None,
-        aliases: Optional[List[str]] = None,
-        distributed_id: Optional[str] = None,
-    ) -> wandb_artifacts.Artifact:
+        artifact_or_path,
+        name = None,
+        type = None,
+        aliases = None,
+        distributed_id = None,
+    ):
         """ Finish a non-finalized artifact as output of a run. Subsequent "upserts" with
         the same distributed ID will result in a new version
 
@@ -1953,13 +1953,13 @@ class Run(object):
 
     def _log_artifact(
         self,
-        artifact_or_path: Union[wandb_artifacts.Artifact, str],
-        name: Optional[str] = None,
-        type: Optional[str] = None,
-        aliases: Optional[List[str]] = None,
-        distributed_id: Optional[str] = None,
-        finalize: bool = True,
-    ) -> wandb_artifacts.Artifact:
+        artifact_or_path,
+        name = None,
+        type = None,
+        aliases = None,
+        distributed_id = None,
+        finalize = True,
+    ):
         if not finalize and distributed_id is None:
             raise TypeError("Must provide distributed_id if artifact is not finalize")
         artifact, aliases = self._prepare_artifact(
@@ -1973,7 +1973,7 @@ class Run(object):
             )
         return artifact
 
-    def _public_api(self) -> PublicApi:
+    def _public_api(self):
         overrides = {"run": self.id}
         run_obj = self._run_obj
         if run_obj is not None:
@@ -1982,7 +1982,7 @@ class Run(object):
         return public.Api(overrides)
 
     # TODO(jhr): annotate this
-    def _assert_can_log_artifact(self, artifact) -> None:  # type: ignore
+    def _assert_can_log_artifact(self, artifact):  # type: ignore
         if not self._settings._offline:
             public_api = self._public_api()
             expected_type = public.Artifact.expected_type(
@@ -2000,11 +2000,11 @@ class Run(object):
 
     def _prepare_artifact(
         self,
-        artifact_or_path: Union[wandb_artifacts.Artifact, str],
-        name: Optional[str] = None,
-        type: Optional[str] = None,
-        aliases: Optional[List[str]] = None,
-    ) -> Tuple[wandb_artifacts.Artifact, List[str]]:
+        artifact_or_path,
+        name = None,
+        type = None,
+        aliases = None,
+    ):
         aliases = aliases or ["latest"]
         if isinstance(artifact_or_path, str):
             if name is None:
@@ -2067,23 +2067,23 @@ class Run(object):
 
     def _set_console(
         self,
-        use_redirect: bool,
-        stdout_slave_fd: Optional[int],
-        stderr_slave_fd: Optional[int],
-    ) -> None:
+        use_redirect,
+        stdout_slave_fd,
+        stderr_slave_fd,
+    ):
         self._use_redirect = use_redirect
         self._stdout_slave_fd = stdout_slave_fd
         self._stderr_slave_fd = stderr_slave_fd
 
-    def __enter__(self) -> "Run":
+    def __enter__(self):
         return self
 
     def __exit__(
         self,
-        exc_type: Type[BaseException],
-        exc_val: BaseException,
-        exc_tb: TracebackType,
-    ) -> bool:
+        exc_type,
+        exc_val,
+        exc_tb,
+    ):
         exit_code = 0 if exc_type is None else 1
         self.finish(exit_code)
         return exc_type is None
@@ -2091,11 +2091,11 @@ class Run(object):
 
 # We define this outside of the run context to support restoring before init
 def restore(
-    name: str,
-    run_path: Optional[str] = None,
-    replace: bool = False,
-    root: Optional[str] = None,
-) -> Union[None, TextIO]:
+    name,
+    run_path = None,
+    replace = False,
+    root = None,
+):
     """ Downloads the specified file from cloud storage into the current directory
         or run directory.  By default this will only download the file if it doesn't
         already exist.
@@ -2158,12 +2158,12 @@ class WriteSerializingFile(object):
     """Wrapper for a file object that serializes writes.
     """
 
-    def __init__(self, f: BinaryIO) -> None:
+    def __init__(self, f):
         self.lock = threading.Lock()
         self.f = f
 
     # TODO(jhr): annotate this
-    def write(self, *args, **kargs) -> None:  # type: ignore
+    def write(self, *args, **kargs):  # type: ignore
         self.lock.acquire()
         try:
             self.f.write(*args, **kargs)
@@ -2171,7 +2171,7 @@ class WriteSerializingFile(object):
         finally:
             self.lock.release()
 
-    def close(self) -> None:
+    def close(self):
         self.lock.acquire()  # wait for pending writes
         try:
             self.f.close()
@@ -2179,7 +2179,7 @@ class WriteSerializingFile(object):
             self.lock.release()
 
 
-def finish(exit_code: int = None) -> None:
+def finish(exit_code = None):
     """
     Marks a run as finished, and finishes uploading all data.
