+++ wandb/sdk_py27/wandb_settings.py
@@ -1,4 +1,4 @@
-#
+# File is generated by: tox -e codemod
 """
 This module configures settings for wandb runs.
 
@@ -57,7 +57,7 @@ if wandb.TYPE_CHECKING:  # type: ignore
         Sequence,
     )
 
-defaults: Dict[str, Union[str, int, Tuple]] = dict(
+defaults = dict(
     base_url="https://api.wandb.ai",
     summary_warnings=5,
     git_remote="origin",
@@ -65,9 +65,9 @@ defaults: Dict[str, Union[str, int, Tuple]] = dict(
 )
 
 # env mapping?
-env_prefix: str = "WANDB_"
+env_prefix = "WANDB_"
 # env_override_suffix: str = "_OVERRIDE"
-env_settings: Dict[str, Optional[str]] = dict(
+env_settings = dict(
     entity=None,
     project=None,
     base_url=None,
@@ -96,12 +96,12 @@ env_settings: Dict[str, Optional[str]] = dict(
     run_job_type="WANDB_JOB_TYPE",
 )
 
-env_convert: Dict[str, Callable[[str], List[str]]] = dict(
+env_convert = dict(
     run_tags=lambda s: s.split(","), ignore_globs=lambda s: s.split(","),
 )
 
 
-def _build_inverse_map(prefix: str, d: Dict[str, str]) -> Dict[str, str]:
+def _build_inverse_map(prefix, d):
     inv_map = dict()
     for k, v in six.iteritems(d):
         v = v or prefix + k.upper()
@@ -109,7 +109,7 @@ def _build_inverse_map(prefix: str, d: Dict[str, str]) -> Dict[str, str]:
     return inv_map
 
 
-def _error_choices(value: str, choices: Set[str]) -> str:
+def _error_choices(value, choices):
     return "{} not in {}".format(value, ",".join(list(choices)))
 
 
@@ -150,7 +150,7 @@ def _get_program_relpath_from_gitrepo(program, _logger=None):
 # the setting exposed to users as `dir=` or `WANDB_DIR` is actually
 # the `root_dir`. We add the `__stage_dir__` to it to get the full
 # `wandb_dir`
-def get_wandb_dir(root_dir: str):
+def get_wandb_dir(root_dir):
     # We use the hidden version if it already exists, otherwise non-hidden.
     if os.path.exists(os.path.join(root_dir, ".wandb")):
         __stage_dir__ = ".wandb" + os.sep
@@ -165,7 +165,7 @@ def get_wandb_dir(root_dir: str):
     return path
 
 
-def _str_as_bool(val) -> Optional[bool]:
+def _str_as_bool(val):
     try:
         val = bool(strtobool(val))
     except (AttributeError, ValueError):
@@ -191,93 +191,93 @@ class Settings(object):
         Exception: if problem.
     """
 
-    mode: str = "online"
-    console: str = "auto"
-    disabled: bool = False
-    run_tags: Optional[Tuple] = None
+    mode = "online"
+    console = "auto"
+    disabled = False
+    run_tags = None
 
-    resume_fname_spec: Optional[str] = None
-    root_dir: Optional[str] = None
-    log_dir_spec: Optional[str] = None
-    log_user_spec: Optional[str] = None
-    log_internal_spec: Optional[str] = None
-    sync_file_spec: Optional[str] = None
-    sync_dir_spec: Optional[str] = None
-    files_dir_spec: Optional[str] = None
-    log_symlink_user_spec: Optional[str] = None
-    log_symlink_internal_spec: Optional[str] = None
-    sync_symlink_latest_spec: Optional[str] = None
-    settings_system_spec: Optional[str] = None
-    settings_workspace_spec: Optional[str] = None
-    silent: bool = False
-    show_info: bool = True
-    show_warnings: bool = True
-    show_errors: bool = True
-    email: Optional[str] = None
-    save_code: Optional[bool] = None
-    program_relpath: Optional[str] = None
+    resume_fname_spec = None
+    root_dir = None
+    log_dir_spec = None
+    log_user_spec = None
+    log_internal_spec = None
+    sync_file_spec = None
+    sync_dir_spec = None
+    files_dir_spec = None
+    log_symlink_user_spec = None
+    log_symlink_internal_spec = None
+    sync_symlink_latest_spec = None
+    settings_system_spec = None
+    settings_workspace_spec = None
+    silent = False
+    show_info = True
+    show_warnings = True
+    show_errors = True
+    email = None
+    save_code = None
+    program_relpath = None
 
     # Public attributes
-    entity: Optional[str] = None
-    project: Optional[str] = None
-    run_group: Optional[str] = None
-    run_name: Optional[str] = None
-    run_notes: Optional[str] = None
-    sagemaker_disable: Optional[bool] = None
+    entity = None
+    project = None
+    run_group = None
+    run_name = None
+    run_notes = None
+    sagemaker_disable = None
 
     # TODO(jhr): Audit these attributes
-    run_job_type: Optional[str] = None
-    base_url: Optional[str] = None
+    run_job_type = None
+    base_url = None
 
     # Private attributes
-    __start_time: Optional[float]
-    __start_datetime: Optional[datetime]
+    # __start_time: Optional[float]
+    # __start_datetime: Optional[datetime]
 
     # Internal attributes
-    __frozen: bool
-    __defaults_dict: Dict[str, int]
-    __override_dict: Dict[str, int]
-    __defaults_dict_set: Dict[str, Set[int]]
-    __override_dict_set: Dict[str, Set[int]]
+    # __frozen: bool
+    # __defaults_dict: Dict[str, int]
+    # __override_dict: Dict[str, int]
+    # __defaults_dict_set: Dict[str, Set[int]]
+    # __override_dict_set: Dict[str, Set[int]]
 
     @enum.unique
     class Source(enum.IntEnum):
-        BASE: int = 1
-        ORG: int = 2
-        ENTITY: int = 3
-        PROJECT: int = 4
-        USER: int = 5
-        SYSTEM: int = 6
-        WORKSPACE: int = 7
-        ENV: int = 8
-        SETUP: int = 9
-        LOGIN: int = 10
-        INIT: int = 11
-        SETTINGS: int = 12
-        ARGS: int = 13
+        BASE = 1
+        ORG = 2
+        ENTITY = 3
+        PROJECT = 4
+        USER = 5
+        SYSTEM = 6
+        WORKSPACE = 7
+        ENV = 8
+        SETUP = 9
+        LOGIN = 10
+        INIT = 11
+        SETTINGS = 12
+        ARGS = 13
 
-    Console: Type[SettingsConsole] = SettingsConsole
+    Console = SettingsConsole
 
     def __init__(  # pylint: disable=unused-argument
         self,
-        base_url: str = None,
-        api_key: str = None,
+        base_url = None,
+        api_key = None,
         anonymous=None,
-        mode: str = None,
-        entity: str = None,
-        project: str = None,
-        run_group: str = None,
-        run_job_type: str = None,
-        run_id: str = None,
-        run_name: str = None,
-        run_notes: str = None,
-        resume: str = None,
-        magic: Union[Dict, str, bool] = False,
-        run_tags: Sequence = None,
+        mode = None,
+        entity = None,
+        project = None,
+        run_group = None,
+        run_job_type = None,
+        run_id = None,
+        run_name = None,
+        run_notes = None,
+        resume = None,
+        magic = False,
+        run_tags = None,
         sweep_id=None,
-        allow_val_change: bool = None,
-        force: bool = None,
-        relogin: bool = None,
+        allow_val_change = None,
+        force = None,
+        relogin = None,
         # compatibility / error handling
         # compat_version=None,  # set to "0.8" for safer defaults for older users
         # strict=None,  # set to "on" to enforce current best practices (also "warn")
@@ -325,7 +325,7 @@ class Settings(object):
         username=None,
         email=None,
         docker=None,
-        sagemaker_disable: Optional[bool] = None,
+        sagemaker_disable = None,
         _start_time=None,
         _start_datetime=None,
         _cli_only_mode=None,  # avoid running any code specific for runs
@@ -378,7 +378,7 @@ class Settings(object):
             self.root_dir = os.path.abspath(os.getcwd())
 
     @property
-    def _offline(self) -> bool:
+    def _offline(self):
         ret = False
         if self.disabled:
             ret = True
@@ -387,53 +387,53 @@ class Settings(object):
         return ret
 
     @property
-    def _silent(self) -> Optional[bool]:
+    def _silent(self):
         if not self.silent:
             return None
         return _str_as_bool(self.silent)
 
     @property
-    def _show_info(self) -> Optional[bool]:
+    def _show_info(self):
         if not self.show_info:
             return None
         return _str_as_bool(self.show_info)
 
     @property
-    def _show_warnings(self) -> Optional[bool]:
+    def _show_warnings(self):
         if not self.show_warnings:
             return None
         return _str_as_bool(self.show_warnings)
 
     @property
-    def _show_errors(self) -> Optional[bool]:
+    def _show_errors(self):
         if not self.show_errors:
             return None
         return _str_as_bool(self.show_errors)
 
     @property
-    def _noop(self) -> bool:
+    def _noop(self):
         return self.mode == "disabled"
 
     @property
-    def _jupyter(self) -> bool:
+    def _jupyter(self):
         return _get_python_type() != "python"
 
     @property
-    def _kaggle(self) -> bool:
+    def _kaggle(self):
         return util._is_kaggle()
 
     @property
-    def _windows(self) -> bool:
+    def _windows(self):
         return platform.system() == "Windows"
 
     @property
-    def _console(self) -> SettingsConsole:
-        convert_dict: Dict[str, SettingsConsole] = dict(
+    def _console(self):
+        convert_dict = dict(
             off=SettingsConsole.OFF,
             wrap=SettingsConsole.WRAP,
             redirect=SettingsConsole.REDIRECT,
         )
-        console: str = self.console
+        console = self.console
         if console == "auto":
             if self._jupyter:
                 console = "wrap"
@@ -462,55 +462,55 @@ class Settings(object):
                 #     console = "redirect"
             else:
                 console = "redirect"
-        convert: SettingsConsole = convert_dict[console]
+        convert = convert_dict[console]
         return convert
 
     @property
-    def resume_fname(self) -> str:
+    def resume_fname(self):
         return self._path_convert(self.resume_fname_spec)
 
     @property
-    def wandb_dir(self) -> str:
+    def wandb_dir(self):
         return get_wandb_dir(self.root_dir or "")
 
     @property
-    def log_user(self) -> str:
+    def log_user(self):
         return self._path_convert(self.log_dir_spec, self.log_user_spec)
 
     @property
-    def log_internal(self) -> str:
+    def log_internal(self):
         return self._path_convert(self.log_dir_spec, self.log_internal_spec)
 
     @property
-    def _sync_dir(self) -> str:
+    def _sync_dir(self):
         return self._path_convert(self.sync_dir_spec)
 
     @property
-    def sync_file(self) -> str:
+    def sync_file(self):
         return self._path_convert(self.sync_dir_spec, self.sync_file_spec)
 
     @property
-    def files_dir(self) -> str:
+    def files_dir(self):
         return self._path_convert(self.files_dir_spec)
 
     @property
-    def log_symlink_user(self) -> str:
+    def log_symlink_user(self):
         return self._path_convert(self.log_symlink_user_spec)
 
     @property
-    def log_symlink_internal(self) -> str:
+    def log_symlink_internal(self):
         return self._path_convert(self.log_symlink_internal_spec)
 
     @property
-    def sync_symlink_latest(self) -> str:
+    def sync_symlink_latest(self):
         return self._path_convert(self.sync_symlink_latest_spec)
 
     @property
-    def settings_system(self) -> str:
+    def settings_system(self):
         return self._path_convert(self.settings_system_spec)
 
     @property
-    def settings_workspace(self) -> str:
+    def settings_workspace(self):
         return self._path_convert(self.settings_workspace_spec)
 
     def _validate_mode(self, value):
@@ -575,8 +575,8 @@ class Settings(object):
         return value
 
     def _start_run(self):
-        datetime_now: datetime = datetime.now()
-        time_now: float = time.time()
+        datetime_now = datetime.now()
+        time_now = time.time()
         object.__setattr__(self, "_Settings__start_datetime", datetime_now)
         object.__setattr__(self, "_Settings__start_time", time_now)
 
@@ -728,10 +728,10 @@ class Settings(object):
         self._update(__d, **kwargs)
 
     def _priority_failed(
-        self, k: str, source: Optional[int], override: Optional[int]
-    ) -> bool:
-        key_source: Optional[int] = self.__defaults_dict.get(k)
-        key_override: Optional[int] = self.__override_dict.get(k)
+        self, k, source, override
+    ):
+        key_source = self.__defaults_dict.get(k)
+        key_override = self.__override_dict.get(k)
         if not key_source or not source:
             return False
         if key_override and not override:
@@ -874,7 +874,7 @@ class Settings(object):
         self.__frozen = True
         return self
 
-    def is_frozen(self) -> bool:
+    def is_frozen(self):
         return self.__frozen
 
     def _load(self, fname):
