+++ wandb/sdk_py27/wandb_setup.py
@@ -1,4 +1,4 @@
-#
+# File is generated by: tox -e codemod
 """Setup wandb session.
 
 This module configures a wandb session which can extend to mutiple wandb runs.
