+++ wandb/sdk_py27/wandb_summary.py
@@ -1,4 +1,4 @@
-#
+# File is generated by: tox -e codemod
 import abc
 
 import six
@@ -27,7 +27,7 @@ class SummaryDict(object):
         raise NotImplementedError
 
     @abc.abstractmethod
-    def _update(self, record: SummaryRecord):
+    def _update(self, record):
         raise NotImplementedError
 
     def keys(self):
@@ -67,7 +67,7 @@ class SummaryDict(object):
 
     __delitem__ = __delattr__
 
-    def update(self, d: t.Dict):
+    def update(self, d):
         # import ipdb; ipdb.set_trace()
         record = SummaryRecord()
         for key, value in six.iteritems(d):
@@ -112,23 +112,23 @@ class Summary(SummaryDict):
         ```
     """
 
-    _update_callback: t.Callable
-    _get_current_summary_callback: t.Callable
+    # _update_callback: t.Callable
+    # _get_current_summary_callback: t.Callable
 
-    def __init__(self, get_current_summary_callback: t.Callable):
+    def __init__(self, get_current_summary_callback):
         super(Summary, self).__init__()
         object.__setattr__(self, "_update_callback", None)
         object.__setattr__(
             self, "_get_current_summary_callback", get_current_summary_callback
         )
 
-    def _set_update_callback(self, update_callback: t.Callable):
+    def _set_update_callback(self, update_callback):
         object.__setattr__(self, "_update_callback", update_callback)
 
     def _as_dict(self):
         return self._get_current_summary_callback()
 
-    def _update(self, record: SummaryRecord):
+    def _update(self, record):
         if self._update_callback:
             self._update_callback(record)
 
@@ -137,9 +137,9 @@ class SummarySubDict(SummaryDict):
     """Non-root node of the summary data structure. Contains a path to itself
     from the root."""
 
-    _items: t.Dict
-    _parent: SummaryDict
-    _parent_key: str
+    # _items: t.Dict
+    # _parent: SummaryDict
+    # _parent_key: str
 
     def __init__(self):
         object.__setattr__(self, "_items", dict())
@@ -149,5 +149,5 @@ class SummarySubDict(SummaryDict):
     def _as_dict(self):
         return self._items
 
-    def _update(self, record: SummaryRecord):
+    def _update(self, record):
         return self._parent._update(record._add_next_parent(self._parent_key))
