+++ wandb/sdk_py27/wandb_watch.py
@@ -1,4 +1,4 @@
-#
+# File is generated by: tox -e codemod
 """
 watch.
 """
